26/12/18 Commit
Continued work on opening pre-existing cards mostly in regards to the keywords the cards will have. Decided to use an external text file and read the list of keywords from that, saves the maintainance of hardcoding them all in seperatly. Am looking at doing the same thing for card types.