29/12/18 Commit
Finally got the basic database functionality finished, so now I can add and remove things from all the tables without issue. Next on the list of functionalities is get the program to render the images from the information it is given. Then styling and optimisation (TODO: Get the SQL queries to clean up after themselves so the program wont lock up if the disc is in heavy use)