30/12/18 Commit
Got the program drawing the card, artwork and textboxes. Also included my own routine for cropping the artwork because the default clipping method defaults to the top right and I want it to be based on the center. Now I just need to get it writting the text and saving the image as opposed to just showing it.