01/01/19 Commit V1.0 Release
It's is finally functionally done, the text renders properly now (the abilities caused some issues with registry as to where they were written) and more importantly you can export the cards in 3 formats Bleed (for printing), Cropped (for printing if you wanna save space and are doing it at home) and Vassal which is for the Vassal card engine or more to the point the Tabletop simulator. The images are saved in subdirectories in a folder you specify and are of 32bit PNG format.