02/01/18
Fixed some bugs regarding to image files not being found.
More importantly ovehauled how the the program looks for information in the database (using the card_code which will be unique not the id coloumn because deleting a card doesnt free its id up so card duplicates with the same id or code could be a thing and that caused issues) This required some major overhauls to how adding and saving cards worked as well as a few other queries to change the "WHERE" clause to refer to a card code which then had to be gotten with something other than a loop. but program is more robust now and if you delete a card it will replace its card code (from a .txt file it generates when you delete a card, which is in and of itself deleted when there are no more missing indexes)